+++ src/roms.hpp
@@ -1,0 +1,7 @@
+extern "C" {
+    extern const unsigned short rom_small_font[1216];
+    extern const unsigned char rom_songlist[7980];
+    extern const unsigned char rom_bgm[822753];
+    extern const unsigned char rom_k8x12S_jisx0201[3072];
+    extern const unsigned char rom_k8x12S_jisx0208[106032];
+}
