+++ src/vgsdecv.hpp
