+++ src/vgssdk.h
