+++ tools/vgsmml/Makefile
@@ -1,4 +1,4 @@
 all: vgsmml
 
 vgsmml: src/vgsmml.c src/vgsmml.h src/cli.c
-	gcc -o vgsmml -I./src -I../../include src/vgsmml.c src/cli.c
+	gcc -o vgsmml -I./src -I../../src src/vgsmml.c src/cli.c
